requirements.txt
discord
asyncio~=3.4.3
PyNaCl #일부만
discord-py-slash-command~=3.0.0
python-dotenv
Neis_API #school_notificator_bot만